backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langchain-community>=0.0.10
pinecone>=2.2.2
tavily-python>=0.2.2
mcp>=0.3.0
loguru>=0.7.0
requests>=2.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
httpx>=0.24.0